metadataを修正
