+++ app/[locale]/errorPlaceholder/page.tsx
@@ -1,8 +1,12 @@
+import { initMetadata, MetadataProps } from "@/metadata.js";
 import { ErrorPage } from "@/common/box.js";
-import { metaDataTitle } from "@/common/title.js";
 
-export const metadata = metaDataTitle(`PLACEHOLDER_TITLE`);
+export async function generateMetadata({ params }: MetadataProps) {
+  return initMetadata(params, "/", "PLACEHOLDER_TITLE");
+}
 
 export default function NotFoundPage() {
-  return <ErrorPage status="PLACEHOLDER_STATUS" message="PLACEHOLDER_MESSAGE" />;;
+  return (
+    <ErrorPage status="PLACEHOLDER_STATUS" message="PLACEHOLDER_MESSAGE" />
+  );
 }
