+++ app/[locale]/main/about/[aboutIndex]/page.tsx
@@ -2,44 +2,44 @@ import { IndexMain } from "@/main/main.js";
 import { AboutContent2 } from "./2-play.js";
 import { AboutContent1 } from "./1-about.js";
 import { AboutContent3 } from "./3-edit.js";
-import { metaDataTitle } from "@/common/title.js";
 import { Pager } from "@/common/pager.js";
 import { AboutContent4 } from "./4-level.js";
 import { AboutContent5 } from "./5-judge.js";
-import { Params } from "next/dist/server/request/params";
+import { getTranslations } from "@/getTranslations.js";
+import { initMetadata } from "@/metadata.js";
 
-const aboutTitles = [
-  "",
-  "概要",
-  "遊び方",
-  "譜面を作ろう",
-  "難易度について",
-  "ゲームの仕様",
-];
 const maxIndex = 5;
 
 export const dynamic = "force-static";
 export const dynamicParams = false;
 export function generateStaticParams() {
-  return Array.from(new Array(maxIndex)).map((_, i) => ({ aboutIndex: (i + 1).toString() }));
+  return Array.from(new Array(maxIndex)).map((_, i) => ({
+    aboutIndex: (i + 1).toString(),
+  }));
 }
 
-export async function generateMetadata(context: { params: Promise<Params> }) {
-  const aboutIndex = Number((await context.params).aboutIndex);
-  return metaDataTitle(aboutTitles[aboutIndex]);
+interface Props {
+  params: Promise<{ locale: string; aboutIndex: string }>;
+}
+export async function generateMetadata({ params }: Props) {
+  const aboutIndex = (await params).aboutIndex;
+  const t = await getTranslations(params, `main.about.${aboutIndex}`);
+  return initMetadata(params, `/main/about/${aboutIndex}`, t("title"));
 }
 
-export default async function AboutTab(context: { params: Promise<Params> }) {
-  const aboutIndex = Number((await context.params).aboutIndex);
+export default async function AboutTab({ params }: Props) {
+  const aboutIndex = Number((await params).aboutIndex);
+  const locale = (await params).locale;
+  const t = await getTranslations(params, `main.about.${aboutIndex}`);
 
   return (
-    <IndexMain tab={0}>
+    <IndexMain tab={0} locale={locale}>
       <Pager
         index={aboutIndex}
         maxIndex={maxIndex}
-        hrefBefore={`/main/about/${aboutIndex - 1}`}
-        hrefAfter={`/main/about/${aboutIndex + 1}`}
-        title={aboutTitles[aboutIndex]}
+        hrefBefore={`/${locale}/main/about/${aboutIndex - 1}`}
+        hrefAfter={`/${locale}/main/about/${aboutIndex + 1}`}
+        title={t("title")}
       />
       <div className="flex-1 text-center break-keep break-words">
         {aboutIndex === 1 ? (
