--- app/[locale]/main/version/layout.tsx
@@ -1,9 +1,0 @@
-import { metaDataTitle } from "@/common/title.js";
-import { tabTitles } from "../const.js";
-import { ReactNode } from "react";
-
-export const metadata = metaDataTitle(tabTitles[4]);
-
-export default function Page(props: {children:ReactNode[]}) {
-    return props.children;
-}
