--- app/[locale]/play/layout.tsx
@@ -1,11 +1,0 @@
-import { ReactNode } from "react";
-
-export const metadata = {
-    openGraph: {
-        url: "/",
-    }
-};
-
-export default function Page(props: { children: ReactNode[] }) {
-    return props.children;
-}
