--- app/[locale]/share/placeholder/layout.tsx
@@ -1,9 +1,0 @@
-import { metaDataTitle } from "@/common/title.js";
-import { ReactNode } from "react";
-
-export const metadata = metaDataTitle("PLACEHOLDER_TITLE");
-// pageTitle(cid, brief) or `Not Found (ID: ${cid})`
-
-export default function Page(props: { children: ReactNode[] }) {
-  return props.children;
-}
